refactor(cozy-realtime): Move main module in its own file
We now have a clean entry point in which we can add exports
